% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamR.R
\name{get_friend_list}
\alias{get_friend_list}
\title{Get Friend List}
\usage{
get_friend_list(api_key, steam_id)
}
\arguments{
\item{api_key}{Your Steam API key.}

\item{steam_id}{The Steam ID of the user.}
}
\value{
A list of friend relationships.
}
\description{
Retrieves the friend list for a specified Steam user. (Note: The user’s friend list must be public.)
}
\examples{
\dontrun{
  get_friend_list("YOUR_API_KEY", "76561198000000000")
}
}
