% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamR.R
\name{get_global_achievement_percentages_for_app}
\alias{get_global_achievement_percentages_for_app}
\title{Get Global Achievement Percentages for a Game}
\usage{
get_global_achievement_percentages_for_app(app_id)
}
\arguments{
\item{app_id}{The application ID of the game.}
}
\value{
A data frame containing achievement percentages.
}
\description{
Retrieves global achievement statistics for a specified game.
}
\examples{
\dontrun{
  get_global_achievement_percentages_for_app("440")
}
}
