% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamR.R
\name{get_owned_games}
\alias{get_owned_games}
\title{Get Owned Games}
\usage{
get_owned_games(api_key, steam_id, include_appinfo = TRUE)
}
\arguments{
\item{api_key}{Your Steam API key.}

\item{steam_id}{The Steam ID of the user.}

\item{include_appinfo}{Logical; if TRUE, includes additional game details.}
}
\value{
A data frame containing the user's owned games.
}
\description{
Retrieves the list of games owned by a Steam user, including playtime data.
}
\examples{
\dontrun{
  get_owned_games("YOUR_API_KEY", "76561198000000000", include_appinfo = TRUE)
}
}
