% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamR.R
\name{get_player_achievements}
\alias{get_player_achievements}
\title{Get Player Achievements}
\usage{
get_player_achievements(api_key, steam_id, app_id, language = "en")
}
\arguments{
\item{api_key}{Your Steam API key.}

\item{steam_id}{The Steam ID of the user.}

\item{app_id}{The application ID of the game.}

\item{language}{The language code for achievement names (default is "en").}
}
\value{
A list of player achievements.
}
\description{
Retrieves the achievements of a Steam user for a specific game.
}
\examples{
\dontrun{
  get_player_achievements("YOUR_API_KEY", "76561198000000000", "440")
}
}
