% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamR.R
\name{get_player_summaries}
\alias{get_player_summaries}
\title{Get Player Summaries}
\usage{
get_player_summaries(api_key, steam_ids)
}
\arguments{
\item{api_key}{Your Steam API key.}

\item{steam_ids}{A character vector of Steam IDs.}
}
\value{
A list of player summaries.
}
\description{
Retrieves basic profile information (e.g., username, avatar, online status) for one or more Steam users.
}
\examples{
\dontrun{
  get_player_summaries("YOUR_API_KEY", c("76561198000000000", "76561198000000001"))
}
}
