% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamR.R
\name{steam_api_request}
\alias{steam_api_request}
\title{Helper Function: steam_api_request}
\usage{
steam_api_request(url, query_params)
}
\arguments{
\item{url}{A character string with the API endpoint URL.}

\item{query_params}{A named list of query parameters.}
}
\value{
A list with the parsed JSON response.
}
\description{
This internal function creates and performs a GET request to the Steam API using httr2.
It adds query parameters, sets a retry policy, and parses the JSON response.
}
\keyword{internal}
